+++ examples/search/select_metrics.py
@@ -4,7 +4,7 @@ from aim.sdk import select_metrics
 if __name__ == '__main__':
     metrics = select_metrics((
         'loss if '
-        '   experiment == test_indicator '
-        '   and run.archived is True '
+        '   run.archived is False '
+        '   and run.loss.train > 0.01 '
     ))
     print(metrics)